requirements.txt
requests==2.28.2
flask==1.1.4
markupsafe==2.0.1
Pillow
opencv-python
torchgeometry
tqdm
tensorboardX
scikit-image
scipy
numpy==1.20